Minor fixes. Documentation Updated
