--- .config/.env
@@ -1,2 +1,0 @@
-export DB_PATH="./tmp/tictactoe.db"
-export DB_URL="sqlite+pysqlite:///./tmp/tictactoe.db"
